Make OperatorEvaluator an enum and eliminate several serving types.
