+++ server/src/main/java/io/spine/server/storage/memory/EntityQueryMatcher.java
@@ -39,7 +39,7 @@ import java.util.Map;
 
 import static com.google.common.base.Preconditions.checkNotNull;
 import static io.spine.protobuf.TypeConverter.toObject;
-import static io.spine.server.storage.memory.QueryOperators.eval;
+import static io.spine.server.storage.memory.OperatorEvaluator.eval;
 import static io.spine.util.Exceptions.newIllegalArgumentException;
 
 /**
