--- server/src/main/java/io/spine/server/storage/memory/QueryOperators.java
@@ -1,81 +1,0 @@
-/*
- * Copyright 2017, TeamDev Ltd. All rights reserved.
- *
- * Redistribution and use in source and/or binary forms, with or without
- * modification, must retain the above copyright notice and the following
- * disclaimer.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
- * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
- * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
- * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
- * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
- * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-package io.spine.server.storage.memory;
-
-import com.google.protobuf.Timestamp;
-import io.spine.client.ColumnFilter.Operator;
-
-import javax.annotation.Nullable;
-import java.util.Objects;
-
-import static com.google.common.base.Preconditions.checkNotNull;
-
-/**
- * A utility class for working with the {@link Operator} enum.
- *
- * <p>The main facility of this class is executing the {@linkplain #eval comparison} operations
- * represented by an {@link Operator} enum value and two operands of a certain type.
- *
- * <a name="supported_types"/>
- *
- * <p>Not all the data types are supported for all the operations:
- * <ol>
- *     <li>{@link Operator#EQUAL EQUAL} supports all the types; the behaviour is equivalent to
- *         {@link Objects#equals(Object, Object)};
- *     <li>other operators currently support only {@link Timestamp} instances comparison; if any of
- *         the operands is {@code null}, the comparison returns {@code false}.
- * </ol>
- *
- * <p>Support for some other data types may be added in future.
- *
- * @author Dmytro Dashenkov
- */
-final class QueryOperators {
-
-    private QueryOperators() {
-        // Prevent this static class from being initialized.
-    }
-
-    /**
-     * Evaluates the given expression.
-     *
-     * <p>For example, if operands where {@code 42} and {@code 9} (exactly in that order) and
-     * the operator was {@link Operator#GREATER_THAN GREATER_THAN}, then this function could be
-     * expressed as {@code 42 > 8}. The function returns the {@code boolean} result of
-     * the evaluation.
-     *
-     * @param left     the left operand
-     * @param operator the comparison operator
-     * @param right    the right operand
-     * @param <T>      the type of the compared values
-     * @return {@code true} if the operands match the operator, {@code false} otherwise
-     * @throws UnsupportedOperationException if the operation is
-     *                                       <a href="supported_types">not supported</a> for
-     *                                       the given data types
-     */
-    static <T> boolean eval(@Nullable T left, Operator operator, @Nullable T right)
-            throws UnsupportedOperationException {
-        checkNotNull(operator);
-        final OperatorEvaluator evaluator = OperatorEvaluators.of(operator);
-        final boolean result = evaluator.eval(left, right);
-        return result;
-    }
-}
