--- server/src/test/java/io/spine/server/storage/memory/OperatorEvaluatorsShould.java
@@ -1,44 +1,0 @@
-/*
- * Copyright 2017, TeamDev Ltd. All rights reserved.
- *
- * Redistribution and use in source and/or binary forms, with or without
- * modification, must retain the above copyright notice and the following
- * disclaimer.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
- * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
- * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
- * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
- * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
- * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-package io.spine.server.storage.memory;
-
-import com.google.common.testing.NullPointerTester;
-import org.junit.Test;
-
-import static com.google.common.testing.NullPointerTester.Visibility.PACKAGE;
-import static io.spine.test.Tests.assertHasPrivateParameterlessCtor;
-
-/**
- * @author Dmytro Dashenkov
- */
-public class OperatorEvaluatorsShould {
-
-    @Test
-    public void have_private_util_ctor() {
-        assertHasPrivateParameterlessCtor(OperatorEvaluators.class);
-    }
-
-    @Test
-    public void not_accept_nulls() {
-        new NullPointerTester()
-                .testStaticMethods(OperatorEvaluators.class, PACKAGE);
-    }
-}
