+++ server/src/test/java/io/spine/server/storage/memory/QueryOperatorShould.java
@@ -20,41 +20,29 @@
 
 package io.spine.server.storage.memory;
 
-import com.google.common.testing.NullPointerTester;
 import com.google.protobuf.Duration;
 import com.google.protobuf.Timestamp;
 import org.junit.Test;
 
 import static com.google.protobuf.util.Timestamps.add;
-import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertTrue;
 import static io.spine.client.ColumnFilter.Operator;
 import static io.spine.client.ColumnFilter.Operator.EQUAL;
 import static io.spine.client.ColumnFilter.Operator.GREATER_OR_EQUAL;
 import static io.spine.client.ColumnFilter.Operator.GREATER_THAN;
 import static io.spine.client.ColumnFilter.Operator.LESS_OR_EQUAL;
 import static io.spine.client.ColumnFilter.Operator.LESS_THAN;
-import static io.spine.server.storage.memory.QueryOperators.eval;
-import static io.spine.test.Tests.assertHasPrivateParameterlessCtor;
+import static io.spine.server.storage.memory.OperatorEvaluator.eval;
 import static io.spine.test.Tests.nullRef;
 import static io.spine.time.Durations2.seconds;
 import static io.spine.time.Time.getCurrentTime;
+import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertTrue;
 
 /**
  * @author Dmytro Dashenkov
  */
 @SuppressWarnings("Duplicates")     // Comparison tests are similar but cannot be simplified to one
-public class QueryOperatorsShould {
-
-    @Test
-    public void have_private_util_ctor() {
-        assertHasPrivateParameterlessCtor(QueryOperators.class);
-    }
-
-    @Test
-    public void not_accept_nulls() {
-        new NullPointerTester().testAllPublicStaticMethods(QueryOperators.class);
-    }
+public class QueryOperatorShould {
 
     @SuppressWarnings("RedundantStringConstructorCall") // We need an equal but not the same object
     @Test
